Refactor Calendar into a container that fetches data and a render component
